Improve safety for constructing LicenseKey objects
Previously one had to manually extract the string corresponding to
the license from a RawLicenseKey object when constructing a
LicenseKey object. That is, the process looked something like:

    LicenseKey::make(raw_license_key->get_license());

However this is both annoying and might lead to unsafe behaviour
since it is not clear that this string based factory does not
check the cryptographic signature (this is done during construction
of the RawLicenseKey object).

The new behaviour is that the make() factory now only accepts
either a RawLicenseKey object or an optional containing a
RawLicenseKey. The old behaviour is available via the make_unsafe()
static factory.