Remove FieldsToReturn struct
The struct was not used anywhere else in the code, and may not be
up to date with current version running on the server. In general
the interface should probably be improved, since the way the
fields to return value is calculated is not compatible with just
ORing together some constants.