Simply interface required for RequestHandlers
Previously we required two methods, one for building the url and
returning it as a string, and one method which made the request
to a certain url.

We change this to only require the method for making the request,
but instead of taking the url we now just pass the method and
parameters like we did for build_url().

In some cases it was unneccesary to build the url as a separate
step, the old way can be accomplished by making build_url() an
internal method in the RequestHandler as shown in this commit.