Reorganize the directory structure
Prevoiusly we most things were in the SKM Client API directory.
This directory added an additional level to the tree without any
clear purpuse, and the spaces in the name being potentially
annoying on unix. Changing to a more standard layout where things
are in the root of the repo. The source is now split among the
include/ and src/ directories.