Remove build_url from interface for RequestHandler
Previously one would make a request in two steps, one would
first call build_url() with the method and a parameter map, then
one would call make_request() with the URL. This two step
process is not neccessary and just complicates e.g. testing and
supporting POST requests and thus we change to just having a single
make_request() method which is called with the method and a
parameter map.