Initial commit of new error handling
Previously errors were only reported through optionals, which made
it hard or impossible to get more detailed information about what
went wrong.

In order to remedy this we add a new class which all methods accept
as the first parameter. In case of an error, more detailed error
information is available in this class after the method call.

Furthermore the design is inspired by monadic error handling and
in case a method is called and the class already has an error set,
this call will turn into a noop. This allows writing several
method calls in a row without having to check for error after every
single call.