Add basic_Error and move constants
We now use the basic_Error class in the library. This class has
been written such that it can be easily subclassed and that no
method names in the class are used, i.e. all methods take a
parameter of type skm::api::main as the first argument leaving
other overloads open for subclasses.