Add cast to make Visual Studio happy. (See desc for details.)
Visual Studio was complaining that the line

    target[tarindex] = (pos - Base64) << 2

resulted in a cast from ptrdiff_t or whatever to unsigned char,
and this might result in loss of data. In this case pos is the
pointer in the Base64 array for the current character, and this
is at most 63, so we are fine.