Start using ResponseParser class
Previously the code for parsing the server responses were part of the
core library. Move this code to the previously added ResponseParser
class. This makes it possible to in the future give the user of the
library an easy way to change which third party libraries are used
for parsing server responses.